Added the admin page.
The admin page allows you to edit existing BSs,
and also technically create your own from their
but it's janky because it results in an error
message even though it works. This can be fixed by
changing the create component so it can allow
multiple different HTTP error codes as successes.

It cannot yet be used to delete existing BSs.

To access this page go to /admin (it's not on the
menu at the bottom of the page though)